refactor: move `RefreshToken`and `TokenStatus` classes to their own files
