+++ lib/core/auth.js
@@ -1,6 +1,7 @@
 import Storage from './storage'
 import { routeOption, isRelativeURL, isSet, isSameURL, getProp } from './utilities'
-import { Token, RefreshToken } from './token'
+import { Token } from './token'
+import { RefreshToken } from './refreshToken'
 
 export default class Auth {
   constructor (ctx, options) {
